# specs/009-notifications/contracts/notifications.openapi.yml
openapi: 3.0.3
info:
  title: KNZiN Notifications API (Feature 009)
  description: Server-authoritative in-app notification center, unread state mutation, user marketing preferences, and administrative broadcast dispatch.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Current API Version

paths:
  /notifications:
    get:
      summary: List authenticated user notifications
      description: Returns a paginated list of notifications for the currently authenticated user in reverse chronological order.
      security:
        - BearerAuth: []
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
            enum: [all, unread]
            default: all
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: Paginated notification collection in JSend format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/NotificationItem'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /notifications/unread-count:
    get:
      summary: Get unread notification badge count
      description: Returns the integer count of unread notifications for the authenticated user to power the header bell badge.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Unread count object in JSend format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      unread_count:
                        type: integer
                        example: 3
        '401':
          $ref: '#/components/responses/Unauthorized'

  /notifications/{id}/read:
    patch:
      summary: Mark a single notification as read
      description: Sets read_at timestamp on the specified notification. Strictly scoped to authenticated user ownership (returns 404 on IDOR).
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      notification:
                        $ref: '#/components/schemas/NotificationItem'
                      unread_count:
                        type: integer
                        example: 2
        '404':
          $ref: '#/components/responses/NotFound'

  /notifications/read-all:
    post:
      summary: Mark all notifications as read
      description: Sets read_at timestamp on all unread notifications belonging to the authenticated user.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All notifications marked read.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      marked_count:
                        type: integer
                        example: 5
                      unread_count:
                        type: integer
                        example: 0

  /notifications/preferences:
    get:
      summary: Get user marketing notification preferences
      description: Retrieves the current toggle state for marketing categories. Transactional notices are not toggleable.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Preferences payload in JSend format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/NotificationPreferences'
    put:
      summary: Update user marketing notification preferences
      description: Updates the category toggles for marketing communications. Disabling a category suppresses both In-App and Email for that category.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                course_announcements:
                  type: boolean
                prize_draw_promotions:
                  type: boolean
                admin_broadcasts:
                  type: boolean
      responses:
        '200':
          description: Updated preferences in JSend format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/NotificationPreferences'

  /notifications/unsubscribe:
    get:
      summary: One-click signed email unsubscribe
      description: Validates cryptographic HMAC signature and updates marketing preferences without requiring interactive login.
      parameters:
        - name: user
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [all, course_announcements, prize_draw_promotions, admin_broadcasts]
            default: all
        - name: signature
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unsubscribe confirmed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Successfully unsubscribed from marketing communications.
        '403':
          description: Invalid or expired signature.

  /admin/notifications/broadcast:
    post:
      summary: Dispatch administrative announcement broadcast
      description: Authorizes an administrator with manage_platform_settings to dispatch platform-wide announcements via in-app and email.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title_ar
                - title_en
                - body_ar
                - body_en
                - channels
              properties:
                title_ar:
                  type: string
                title_en:
                  type: string
                body_ar:
                  type: string
                body_en:
                  type: string
                channels:
                  type: array
                  items:
                    type: string
                    enum: [in_app, email]
      responses:
        '201':
          description: Broadcast queued successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      broadcast_id:
                        type: string
                        format: uuid
                      recipients_queued:
                        type: integer
                        example: 1420
        '403':
          description: Forbidden - Requires manage_platform_settings capability.

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum

  schemas:
    NotificationItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 9d020d20-43aa-448f-9a99-0f2c4189e472
        category:
          type: string
          enum: [transactional, course_announcements, prize_draw_promotions, admin_broadcasts]
          example: transactional
        title:
          type: string
          example: تم استلام طلبك بنجاح
        body:
          type: string
          example: تم تأكيد طلبك رقم KNZ-948171 وحجز 5 تذاكر ترويجية.
        action_type:
          type: string
          enum: [navigate, refresh_course]
          default: navigate
          example: refresh_course
        action_url:
          type: string
          nullable: true
          example: /lessons/electrical-wiring
        entity_type:
          type: string
          nullable: true
          example: course
        entity_id:
          type: string
          nullable: true
          example: 9d020d20-43aa-448f-9a99-0f2c4189e472
        metadata:
          type: object
          nullable: true
          properties:
            course_slug:
              type: string
              example: electrical-wiring
            content_version:
              type: integer
              example: 2
        is_read:
          type: boolean
          example: false
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          example: 2026-10-04T18:00:00Z

    NotificationPreferences:
      type: object
      properties:
        course_announcements:
          type: boolean
          example: true
        prize_draw_promotions:
          type: boolean
          example: true
        admin_broadcasts:
          type: boolean
          example: true
        is_unsubscribed_from_all:
          type: boolean
          example: false

    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 4
        per_page:
          type: integer
          example: 15
        total:
          type: integer
          example: 52

  responses:
    Unauthorized:
      description: Unauthenticated request.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: fail
              message:
                type: string
                example: Unauthenticated.
    NotFound:
      description: Notification not found or not owned by user.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: fail
              message:
                type: string
                example: Notification not found.